// frontend/src/TVShow/Details/TVShowDetails.css.d.ts
interface CSSModule {
  readonly container: string;
  readonly header: string;
  readonly title: string;
  readonly details: string;
  readonly detailRow: string;
  readonly label: string;
  readonly value: string;
  readonly overview: string;
  readonly monitoredIcon: string;
  readonly genres: string;
  readonly genre: string;
}

declare const styles: CSSModule;

export default styles;